el sistema de subida de exámenes vuelve a funcionar plenamente

+++ php/uploader.php
@@ -1,6 +1,4 @@
 <?php
-require 'vendor/autoload.php';
-
 function getTags($tipo_tag, $db){
 	$tags = array();
 	$sql = "SELECT nom_tag FROM Tag WHERE tipo_tag=? GROUP BY nom_tag";
@@ -13,53 +11,25 @@ function getTags($tipo_tag, $db){
 	return $tags;
 }
 
-function getSiglas($tags){
-	$pattern = '/[A-Z]|Á/';
-	foreach($tags as $label){
-		$sigla="";
-		if( preg_match_all($pattern, $label, $matches) ){
-			if( count($matches[0]) >1){
-				foreach ($matches[0] as $val) {
-					if($val == 'Á'){
-						$sigla.='A';
-					}else{
-						$sigla.=$val;
-					}
-				}
-				$siglas[$label] = $sigla;
-			}
-		}
-	}
-	return $siglas;
-}
-
 $db = new PDO("sqlite:../examenes.db");
+$grados = getTags("grado", $db);
+$grados[] = "Informática y Doble Grado";
 $asigs = getTags("asig", $db);
-$asigSiglas = getSiglas($asigs);
 for($i=2011; $i<date("Y"); $i++) {
 	$anios[] = $i;
 }
-$patterns = array('/\s/','/á/','/é/','/í/','/ó/','/ú/');
-$replacements = array('\s+','(a|á)','(e|é)','(i|í)','(o|ó)','(u|ú)');
-$updir='../sinClasificar';
-$ds=DIRECTORY_SEPARATOR;
-
-// print_r($_FILES);
-$filename = $_FILES['file']['name'];
-// foreach($_FILES['file']['name'] as $index=>$filename){
 $md5 = md5_file($_FILES['file']['tmp_name']);
 $sql='SELECT COUNT(*) FROM Documento WHERE id_doc=?';
 $query = $db->prepare($sql);
 $query->execute([$md5]);
-$file = $updir.$ds.$filename;
-$txt="";
+$file = '../sinClasificar/'.$_FILES['file']['name'];
 // Check if file is not uploaded and if it is not in the db
-if($query->fetchColumn()==0){
-	// if(!file_exists($file) and $query->fetchColumn()==0){
+if($query->fetchColumn()==0 and !file_exists($file)){
     move_uploaded_file($_FILES['file']['tmp_name'], $file);
 	$data['asig'] = $asigs;
+	$data['grado'] = $grados;
 	foreach ($anios as $value) {
-		$data['anio'][] = $value-2000;
+		$data['anio'][] = ($value-2000)*100 + $value-1999; // $value=2013, 1314
 	}
 }else{
 	$data['duplicate'] = TRUE;
